Entity beans e os DAOs de Centro de custo, dept, empresa, meta e seus relacionamentos. Recriados apos reuniao de sexta passada

+++ GestorWeb/src/main/java/com/saax/gestorweb/dao/UsuarioDAO.java
@@ -1,26 +1,30 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package com.saax.gestorweb.dao;
 
+import com.saax.gestorweb.dao.exceptions.IllegalOrphanException;
 import com.saax.gestorweb.dao.exceptions.NonexistentEntityException;
 import com.saax.gestorweb.model.datamodel.Usuario;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import com.saax.gestorweb.model.datamodel.UsuarioEmpresa;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 
 /**
- *
- * @author Rodrigo
+ * DAO para o entity bean: Usuario <br><br>
+ * 
+ * Classe gerada automaticamente pelo netbeans: NÃO ALTERAR<br>
+ * 
+ * ATENÇÃO: Não instanciar esta classe diretamente, mas sim sua subclasse substituta: UsuarioDAOCustom
+ * 
+ * @author rodrigo
  */
-public class UsuarioDAO implements Serializable {
+class UsuarioDAO implements Serializable {
 
     public UsuarioDAO(EntityManagerFactory emf) {
         this.emf = emf;
@@ -32,11 +36,29 @@ public class UsuarioDAO implements Serializable {
     }
 
     public void create(Usuario usuario) {
+        if (usuario.getEmpresas() == null) {
+            usuario.setEmpresas(new ArrayList<UsuarioEmpresa>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Collection<UsuarioEmpresa> attachedEmpresas = new ArrayList<UsuarioEmpresa>();
+            for (UsuarioEmpresa empresasUsuarioEmpresaToAttach : usuario.getEmpresas()) {
+                empresasUsuarioEmpresaToAttach = em.getReference(empresasUsuarioEmpresaToAttach.getClass(), empresasUsuarioEmpresaToAttach.getId());
+                attachedEmpresas.add(empresasUsuarioEmpresaToAttach);
+            }
+            usuario.setEmpresas(attachedEmpresas);
             em.persist(usuario);
+            for (UsuarioEmpresa empresasUsuarioEmpresa : usuario.getEmpresas()) {
+                Usuario oldUsuarioOfEmpresasUsuarioEmpresa = empresasUsuarioEmpresa.getUsuario();
+                empresasUsuarioEmpresa.setUsuario(usuario);
+                empresasUsuarioEmpresa = em.merge(empresasUsuarioEmpresa);
+                if (oldUsuarioOfEmpresasUsuarioEmpresa != null) {
+                    oldUsuarioOfEmpresasUsuarioEmpresa.getEmpresas().remove(empresasUsuarioEmpresa);
+                    oldUsuarioOfEmpresasUsuarioEmpresa = em.merge(oldUsuarioOfEmpresasUsuarioEmpresa);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -45,17 +67,50 @@ public class UsuarioDAO implements Serializable {
         }
     }
 
-    public void edit(Usuario usuario) throws NonexistentEntityException, Exception {
+    public void edit(Usuario usuario) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Usuario persistentUsuario = em.find(Usuario.class, usuario.getId());
+            Collection<UsuarioEmpresa> empresasOld = persistentUsuario.getEmpresas();
+            Collection<UsuarioEmpresa> empresasNew = usuario.getEmpresas();
+            List<String> illegalOrphanMessages = null;
+            for (UsuarioEmpresa empresasOldUsuarioEmpresa : empresasOld) {
+                if (!empresasNew.contains(empresasOldUsuarioEmpresa)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain UsuarioEmpresa " + empresasOldUsuarioEmpresa + " since its usuario field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            Collection<UsuarioEmpresa> attachedEmpresasNew = new ArrayList<UsuarioEmpresa>();
+            for (UsuarioEmpresa empresasNewUsuarioEmpresaToAttach : empresasNew) {
+                empresasNewUsuarioEmpresaToAttach = em.getReference(empresasNewUsuarioEmpresaToAttach.getClass(), empresasNewUsuarioEmpresaToAttach.getId());
+                attachedEmpresasNew.add(empresasNewUsuarioEmpresaToAttach);
+            }
+            empresasNew = attachedEmpresasNew;
+            usuario.setEmpresas(empresasNew);
             usuario = em.merge(usuario);
+            for (UsuarioEmpresa empresasNewUsuarioEmpresa : empresasNew) {
+                if (!empresasOld.contains(empresasNewUsuarioEmpresa)) {
+                    Usuario oldUsuarioOfEmpresasNewUsuarioEmpresa = empresasNewUsuarioEmpresa.getUsuario();
+                    empresasNewUsuarioEmpresa.setUsuario(usuario);
+                    empresasNewUsuarioEmpresa = em.merge(empresasNewUsuarioEmpresa);
+                    if (oldUsuarioOfEmpresasNewUsuarioEmpresa != null && !oldUsuarioOfEmpresasNewUsuarioEmpresa.equals(usuario)) {
+                        oldUsuarioOfEmpresasNewUsuarioEmpresa.getEmpresas().remove(empresasNewUsuarioEmpresa);
+                        oldUsuarioOfEmpresasNewUsuarioEmpresa = em.merge(oldUsuarioOfEmpresasNewUsuarioEmpresa);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                Integer id = usuario.getIdUsuario();
+                Integer id = usuario.getId();
                 if (findUsuario(id) == null) {
                     throw new NonexistentEntityException("The usuario with id " + id + " no longer exists.");
                 }
@@ -68,7 +123,7 @@ public class UsuarioDAO implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -76,9 +131,20 @@ public class UsuarioDAO implements Serializable {
             Usuario usuario;
             try {
                 usuario = em.getReference(Usuario.class, id);
-                usuario.getIdUsuario();
+                usuario.getId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The usuario with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            Collection<UsuarioEmpresa> empresasOrphanCheck = usuario.getEmpresas();
+            for (UsuarioEmpresa empresasOrphanCheckUsuarioEmpresa : empresasOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Usuario (" + usuario + ") cannot be destroyed since the UsuarioEmpresa " + empresasOrphanCheckUsuarioEmpresa + " in its empresas field has a non-nullable usuario field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(usuario);
             em.getTransaction().commit();
@@ -134,24 +200,5 @@ public class UsuarioDAO implements Serializable {
             em.close();
         }
     }
-
-    public Usuario findUsuarioByLogin(String login) {
-        EntityManager em = getEntityManager();
-        try {
-
-            Query queryProductsByName = em.createNamedQuery("Usuario.findByLogin");
-            queryProductsByName.setParameter("login", login);
-
-            List results = queryProductsByName.getResultList();
-
-            if (!results.isEmpty()) {
-                // ignores multiple results
-                return (Usuario) results.get(0);
-            }
-            return null;
-        } finally {
-            em.close();
-        }
-    }
-
+    
 }
